Draw interface changes
Simple draw method now only accepts rectangles. This rectangle defines
both the position and the shape.

The exhaustive draw changed its interpretation of the first parameter.
Now it will see it as a shape definition only. This affects how
transformations are applied. Previously, there was some position
shifting involved to ensure transformations are applied to centered
shapes. Now this responsibility is pushed to the caller side.